Extract two services from socketController
